Add log-spaced sparse event handling
Sparse event handling means that every trigger event is registered but the data for some events is skipped. This allows for uses like keeping all the acquired data in memory to decrease the time between trigger events while not having to wait for data to be saved in a per-event basis that may not be scalable.