Add toggle for scaling values before saving to file
Scaling the values is more convenient (as the raw data is then directly in V), but takes 4 times as much memory/storage. The physical values can always be calculated post-hoc, and the record is maintained as long as the corresponding Labber output file is kept (which should be done anyway on account of all the other instrument settings).