Change hdf5 output to save only 1 point
It's all zeroes anyway (as the data is stored in the custom files), so we might as well save space by only having a single-point array in the hdf5 file. I think this is better than having no data points, as then it might not play nicely with some of the Labber functionality.